% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_divisions}
\alias{census_divisions}
\title{census_divisions}
\format{
A \code{data frame} with 293 observations and 6 variables.
}
\source{
Adapted from official Canadian Census shapefiles.
}
\usage{
census_divisions
}
\description{
Geometries for each Census Division (CD)
}
\section{Variables}{


\itemize{
 \item \code{cduid}: Census division number.
 \item \code{cdname}: Census division name.
 \item \code{cdname}: Census division type (i.e. see
  the README in the GitHub repository).
 \item \code{pruid}: Province number.
 \item \code{prname}: Province name.
 \item \code{geometry}: Census division shape.
}
}

\keyword{datasets}
