% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{federal_electoral_districts}
\alias{federal_electoral_districts}
\title{federal_electoral_districts}
\format{
A \code{data frame} with 338 observations and 5 variables.
}
\source{
Adapted from official Canadian Census shapefiles.
}
\usage{
federal_electoral_districts
}
\description{
Geometries for each Federal Electoral District (FED)
}
\section{Variables}{


\itemize{
 \item \code{feduid}: Census division number.
 \item \code{fedname}: Federal electoral district name.
 \item \code{pruid}: Province number.
 \item \code{prname}: Province name.
 \item \code{geometry}: Federal electoral district shape.
}
}

\keyword{datasets}
