% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_economic_zones.R
\name{get_economic_zones}
\alias{get_economic_zones}
\title{Mapa de Canada a nivel zona economica - WRITE IN ENGLISH !!!!}
\usage{
get_economic_zones(map = census_divisions)
}
\arguments{
\item{mapa}{mapa a agregar, por defecto es todo el mapa nacional}
}
\value{
Un objeto de clase sf y data.frame.
}
\description{
Este mapa es calculado a partir
del mapa censal para no recargar el volumen de datos del paquete.
}
\examples{
get_provinces()
}
