% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provinces.R
\name{get_provinces}
\alias{get_provinces}
\title{Mapa de Canada a nivel provincial - WRITE IN ENGLISH !!!!}
\usage{
get_provinces(map = canadamaps::census_divisions)
}
\arguments{
\item{mapa}{mapa a agregar, por defecto es todo el mapa nacional}
}
\value{
Un objeto de clase sf y data.frame.
}
\description{
Este mapa es calculado a partir
del mapa censal para no recargar el volumen de datos del paquete.
}
\examples{
get_provinces()
}
