% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambert_projection.R
\name{lambert_projection}
\alias{lambert_projection}
\title{Project an sf object to the Lambert Conformal Conic projection}
\usage{
lambert_projection(tbl, crs_string = NULL)
}
\arguments{
\item{tbl}{an sf object}

\item{crs_string}{a character string specifying the projection}
}
\value{
an sf object
}
\description{
Project an sf object to the Lambert Conformal Conic projection
}
